xen/arm: Enable the existing MSI-X and MSI handlers support for ARM
The existing MSI-X and MSI handlers support available for X86 is adapted
for Arm.

Enable existing MSIX handlers for accesses to the MSI-X message control
field on the PCI configuration space, and traps for accesses to the
memory region that contains the MSI-X table and PBA for ARM. This traps
detect attempts from the guest to configure MSI-X interrupts and
properly sets them up.

Enable existing handlers for the MSI control, address, data and mask
fields in order to detect accesses to them and setup the interrupts as
requested by the guest.

Signed-off-by: Rahul Singh <rahul.singh@arm.com>
Signed-off-by: Stewart Hildebrand <stewart.hildebrand@amd.com> # rebase
(cherry picked from commit a9736ab31ceec843d75cd870e40ff82c7af1b79a)
---
Stewart: rebase on top of 4.17.0-rc3
* updated section of code moved from xen/drivers/vpci/msix.c to xen/drivers/vpci/x86/msix.c
* s/guest_addr/guest_reg/ in xen/drivers/vpci/arm/msix.c
* add const struct domain *d arg to vpci_msix_arch_print in xen/drivers/vpci/arm/vmsi.c

Note: the referenced commit a9736ab31cee exists in the downstream branch
https://gitlab.com/xen-project/fusa/xen-integration/-/tree/integration/pci-passthrough/

Note that I have not tested this with MSI/MSI-X, GICv3, or x86.