xen/arm: account IO handlers for emulated PCI MSI-X
At the moment, we always allocate an extra 16 slots for IO handlers
(see MAX_IO_HANDLER). So while adding IO trap handlers for the emulated
MSI-X registers we need to explicitly tell that we have additional IO
handlers, so those are accounted.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>
Acked-by: Julien Grall <jgrall@amazon.com>
---
Reviewed-by: Michal Orzel <michal.orzel@amd.com> # xlnx downstream
---
Cc: Julien Grall <julien@xen.org>
Cc: Stefano Stabellini <sstabellini@kernel.org>
---
This actually moved here from the part 2 of the prep work for PCI
passthrough on Arm as it seems to be the proper place for it.

Since v5:
- optimize with IS_ENABLED(CONFIG_HAS_PCI_MSI) since VPCI_MAX_VIRT_DEV is
  defined unconditionally
New in v5
---
(cherry picked from commit af28d4673464d3daac23749547676ed8f528e0d1 from
 the downstream branch virtio_devel_next from
 https://github.com/otyshchenko1/xen.git)

Discussion on list:
https://lists.xenproject.org/archives/html/xen-devel/2022-07/msg01129.html

v2->v3 (xlnx):
* Add R-b