xen/arm: pci device passthrough for dom0less guests
Introduce the xen,pci-assigned property for the passthrough
node that is used to assign pci devices to a dom0less guest.

When the property is found, the addressed pci device is
assigned to the guest and the vpci DT node is added to the
guest DT.

Update the documentation about this property.

Issue-Id: SCM-1012
Signed-off-by: Luca Fancellu <luca.fancellu@arm.com>
Change-Id: I568aea739c1731f6d925a0bfbe581581187e6301
---
(cherry picked from commit b53ea0b0668837870b718ba8445862befcb08663 from
 the downstream branch poc/pci-passthrough from
 https://gitlab.com/xen-project/people/bmarquis/xen-arm-poc.git)

changes since cherry-pick:
* coding style
* add note about devfn
* capitalize U in domU_assign_pci_device
* rebase on ("Revert "xen/arm: Add cmdline boot option "pci-passthrough = <boolean>""")

Signed-off-by: Stewart Hildebrand <stewart.hildebrand@amd.com>