xen/arm: don't pass iommu properties to hwdom for iommu-map
A DMA capable device, such as a PCIe root controller, may have an iommu-map
property with a phandle reference to the SMMU node, but not an iommus property.
In this case, we want to treat it the same as we currently handle devices with
an iommus property: don't pass the iommu related properties to hwdom.

Reported-by: Michal Orzel <michal.orzel@amd.com>
Signed-off-by: Stewart Hildebrand <stewart.hildebrand@amd.com>
---
v4->v5:
* new patch