xen/pci: initialize BARs
A PCI device must have valid BARs in order to assign it to a dom0less
domU. On ARM, firmware is unlikely to have initialized the BARs, so we
must do this in Xen. During pci_init(), check if each BAR is valid. If
the BAR happens to already be valid, add the BAR to a rangeset. If not
valid, reserve a new BAR address from the rangeset and write it to the
device.

Signed-off-by: Stewart Hildebrand <stewart.hildebrand@amd.com>