xen/pci: initialize BARs
A PCI device must have valid BARs in order to assign it to a dom0less
domU. On ARM, firmware is unlikely to have initialized the BARs, so we
must do this in Xen. During pci_init(), check if each BAR is valid. If
the BAR happens to already be valid, remove the BAR range from a
rangeset of valid PCI ranges so as to avoid overlap when reserving a new
BAR. If not valid, reserve a new BAR address from the rangeset and write
it to the device.

Signed-off-by: Stewart Hildebrand <stewart.hildebrand@amd.com>
---
Acked-by: Stefano Stabellini <stefano.stabellini@amd.com> # xlnx

2026.1 v2:
* propagate errors
* zero-check addr
* add prefetch bit during BAR write

2025.1:
* skip I/O BARs in bars_iterate to avoid triggering an ASSERT in
  pci_size_mem_bar()

2024.1 v2:
* move pci_fixup_bars() call
* disable memory / I/O decoding before writing new BAR