Revert "xen/arm: vpci: permit access to guest vpci space"
Since the iomem_access_permitted() check in ("vpci/header: program p2m
with guest BAR view") was changed to use MFNs (it used GFNs in an
earlier rev) this whole patch should be dropped. The toolstack already
does what this patch was trying to do with XEN_DOMCTL_iomem_permission.

This reverts commit 11dd6f208815242455cd7b819f18d12c725ccf31.