arm/smmu: adjust pdev->domain PCI assignment
When assigning a pdev to dom0, pdev->domain may have already been set
elsewhere. However, the SMMU assignment may not have finished
completely. In this case, we need to continue the SMMU assignment to
avoid SMMU faults. Drop the pdev->domain == d check.

Fixes: ("arm/smmu: remove is_hardware_domain check in assign")
Signed-off-by: Stewart Hildebrand <stewart.hildebrand@amd.com>
---
Should I create a CR for this?