vpci/msix: Expose complete MSI-X config space to domUs
The vpci MSI-X handler was only exposing the first 4 of 12 bytes of the
MSI-X config space.  Dom0 must fall back to default HW reading.  For a
domU though this meant the PBA and MSI-X tables were returning 0xff.
Since these values are invalid, the domU wouldn't try to use MSI-X.

Latch the values and return them for a domU.  The fields are read-only
and static, so they can be returned without consulting hardware.

Signed-off-by: Jason Andryuk <jason.andryuk@amd.com>
Signed-off-by: Stewart Hildebrand <stewart.hildebrand@amd.com>
---
This patch was borrowed from VAS and introduced in xlnx_rebase_4.20.
Link to gerrit reviews:
https://gerrit-git.amd.com/c/brahma/ec/xen/+/1146703
https://gerrit-git.amd.com/c/brahma/ec/xen/+/1238864

Since then, I've reworked the cleanup path slightly.